Actually update the users' points
The `update_metadata()` function doesn’t do bulk updates for multiple
users, so we have to get a list of users, and loop through them,
calling `update_user_meta()` for each.

Fixes it so that the points actually get reset. :)

We could have used `wordpoints_set_points()`, but we weren’t logging
this before, so I decided not to for now.